BUG_05 wyslanie maili do wszystkich nie dziala -> route moze zostać tak jak jest powinna być komeda dopisana do script.py
BUG_07 dodawanie dań z koziołka czasami łączą się dania (startswtith ?), nie dododało raz zupy (duża litera ?)

BUG_09 Jak dla mnie Standard Menu ma niepotrzebny left margin. Wyglada to nieładnie, bo jest nierówno z Daily menu. (Chodzi mi o te szare pola)

BUG_14 automatyczne dodawanie dan z koziolka i tomasa - zadanie dla crona
