+++ src/lunch_app/templates/add_food.html
@@ -62,7 +62,7 @@
         To add few meal at one time separate them with enter and click bulk,
         all meals will be added with same company, cost, dates and type.
 <hr><h4>If something gone wrong</h4>
-        To add meal of a day from "Pod Koziołek" click <a href="{{ url_for("add_daily_koziolek") }}">here</a>.
+        To add meal of a day from "Pod Koziołek" click <a href="{{ url_for("add_daily_koziolek_view") }}">here</a>.
 <br><br>
     !!! WARNING: use ONLY on MODAY !!!<br>
     To add a weak from "Tomas" click <a href="{{ url_for("get_week_from_tomas_view") }}">here</a>.<br>
