+++ src/lunch_app/tests.py
@@ -618,6 +618,7 @@ class LunchBackendViewsTestCase(unittest.TestCase):
                     user_order.description,
                 )
 
+    @patch('lunch_app.permissions.current_user', new=MOCK_ADMIN)
     @patch('lunch_app.views.current_user', new=MOCK_ADMIN)
     def test_send_daily_reminder(self):
         """
@@ -832,7 +833,7 @@ class LunchBackendViewsTestCase(unittest.TestCase):
     @patch('lunch_app.views.current_user', new=MOCK_ADMIN)
     @patch('lunch_app.permissions.current_user', new=MOCK_ADMIN)
     @patch(
-        'lunch_app.views.get_dania_dnia_from_pod_koziolek',
+        'lunch_app.utils.get_dania_dnia_from_pod_koziolek',
         new=MOCK_DATA_KOZIOLEK,
     )
     def test_add_daily_koziolek(self):
@@ -859,7 +860,7 @@ class LunchBackendViewsTestCase(unittest.TestCase):
     @patch('lunch_app.views.current_user', new=MOCK_ADMIN)
     @patch('lunch_app.permissions.current_user', new=MOCK_ADMIN)
     @patch(
-        'lunch_app.views.get_week_from_tomas',
+        'lunch_app.utils.get_week_from_tomas',
         new=MOCK_DATA_TOMAS,
     )
     def test_get_week_from_tomas_view(self):
@@ -1110,8 +1111,8 @@ class LunchWebCrawlersTestCases(unittest.TestCase):
         """
         data = get_dania_dnia_from_pod_koziolek()
         self.assertGreaterEqual(len(data), 2)
-        self.assertGreaterEqual(len(data["zupa_dnia"]), 1)
-        self.assertGreaterEqual(len(data['danie_dania_1']), 1)
+        self.assertGreaterEqual(len(data["zupy"]), 1)
+        self.assertGreaterEqual(len(data['dania_dnia']), 1)
 
     @patch(
         'lunch_app.webcrawler.read_webpage',
